software_accounting/requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==21.2.0
marshmallow==3.20.1
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
